Contenedor_Docker/app/requirements.txt
fastapi==0.116.1
psycopg2-binary==2.9.10
uvicorn==0.35.0